src/config/settings.py
from functools import lru_cache

from dotenv import find_dotenv, load_dotenv
from pydantic_settings import BaseSettings, SettingsConfigDict

load_dotenv(find_dotenv(usecwd=True), override=False)


class Settings(BaseSettings):
    app_name: str = "python-claw-gateway"
    database_url: str = "sqlite:///./python_claw.db"
    dedupe_retention_days: int = 30
    dedupe_stale_after_seconds: int = 300
    messages_page_default_limit: int = 50
    messages_page_max_limit: int = 100

    model_config = SettingsConfigDict(
        env_prefix="PYTHON_CLAW_",
        extra="ignore",
    )


@lru_cache
def get_settings() -> Settings:
    return Settings()
