pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-claw"
version = "0.1.0"
description = "Gateway-first OpenClaw-style assistant foundation in Python."
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14,<2.0",
  "fastapi>=0.115,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "python-dotenv>=1.0,<2.0",
  "sqlalchemy>=2.0.36,<3.0",
  "uvicorn>=0.32,<1.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.3,<9.0",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "src*"]
